Most of the things are done :tada:

+++ src/main/java/com/helloworld/repository/AdministratorRepo.java
@@ -3,7 +3,7 @@ package com.helloworld.repository;
 import com.helloworld.model.Administrator;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdministratorRepo extends JpaRepository<Administrator, Integer> {
+public interface AdministratorRepo extends JpaRepository<Administrator, Long> {
 
     Administrator findByEmail(String email);
