+++ src/main/java/com/helloworld/repository/LibrarianRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.helloworld.model.Librarian;
 
-public interface LibrarianRepo extends JpaRepository<Librarian, Integer> {
+public interface LibrarianRepo extends JpaRepository<Librarian, Long> {
 
     Librarian findByEmail(String email);
